feat: Cambio de interfaz grafica
