--- templates/graph.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Grafo de las Tuberías</title>
-    <link rel="stylesheet" href="../static/styles.css">
-</head>
-<body>
-    <h1>Grafo de las Tuberías</h1>
-    <img src="/static/graph.png" alt="Graph">
-    <a href="/templates/index.html" class="btn">Volver al Inicio</a>
-</body>
-</html>
