--- templates/index.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Trabajo Final de Algoritmos</title>
-    <link rel="stylesheet" href="../static/styles.css">
-</head>
-<body>
-    <div class="container">
-        <h1>Trabajo Final de Algoritmos</h1>
-        <div class="buttons">
-            <a href="/templates/graph.html" class="btn">Observar grafo de las tuberias</a>
-        </div>
-        
-        <h2>Calcular Camino Más Corto</h2>
-        <form>
-            <label for="start_node">Nodo de Inicio:</label>
-            <input type="number" id="start_node" name="start_node" min="1" max="1500" required>
-            <label for="end_node">Nodo de Fin:</label>
-            <input type="number" id="end_node" name="end_node" min="1" max="1500" required>
-            <div>
-              <button type="button" class="btn" onclick="window.location.href='/templates/graph.html'">Calcular</button>
-            </div>
-        </form>
-    </div>
-    <div class="container-footer">
-        <div class="footer">
-            <div class="copyright">
-                © 2024 All Rights Reserved | <a>Grupo 07</a>
-            </div>
-        </div>
-    </div>
-    
-</body>
-</html>
